solution/app/requirements.txt
# CymbalFlix Discover - Python Dependencies
# ==========================================

# Web Framework
streamlit>=1.28.0

# Database Connectivity
# AlloyDB Python Connector handles secure IAM-authenticated connections
google-cloud-alloydb-connector[pg8000]>=1.0.0

# SQLAlchemy for connection pooling and query execution
sqlalchemy>=2.0.0

# Data Processing
pandas>=2.0.0

# Charts and Visualization
plotly>=5.18.0

# Environment Variable Management
python-dotenv>=1.0.0
